task3 hw332 change Authentication activity, and chose new start activity
