+++ app/src/main/AndroidManifest.xml
@@ -21,8 +21,8 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".AboutActivity"
-            android:label="@string/about_title">
+        <activity android:name=".InputActivity"
+            android:label="@string/app_name">
         </activity>
     </application>
