--- app/src/main/java/com/example/photoproj/AboutActivity.java
@@ -1,17 +1,0 @@
-package com.example.photoproj;
-
-import android.app.Activity;
-import android.os.Bundle;
-
-/**
- * Created by Alexander Klimov on 01.12.2014.
- */
-public class AboutActivity extends Activity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        setContentView(R.layout.activity_about);
-    }
-}
