+++ app/src/main/java/com/example/photoproj/ColorSchemeSelection.java
@@ -3,6 +3,9 @@ package com.example.photoproj;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
+/**
+ * класс, для описания логики меню смены цветовой схемы приложения
+ */
 public class ColorSchemeSelection extends AppCompatActivity {
 
     @Override
