--- app/src/main/res/layout/activity_about.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent" android:layout_height="match_parent">
-
-    <TextView
-        android:id="@+id/textView_about_content"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Вошли."/>
-
-</ScrollView>
