+++ app/src/main/res/layout/activity_input.xml
@@ -1,0 +1,55 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+xmlns:app="http://schemas.android.com/apk/res-auto"
+xmlns:tools="http://schemas.android.com/tools"
+android:layout_width="match_parent"
+android:layout_height="match_parent"
+app:layout_behavior="@string/appbar_scrolling_view_behavior"
+tools:context=".MainActivity"
+tools:showIn="@layout/app_bar_main">
+
+<android.support.design.widget.TextInputLayout
+    android:id="@+id/textInputLayout"
+    android:layout_width="207dp"
+    android:layout_height="57dp"
+    android:layout_marginStart="8dp"
+    android:layout_marginLeft="8dp"
+    android:layout_marginTop="8dp"
+    android:layout_marginEnd="8dp"
+    android:layout_marginRight="8dp"
+    android:layout_marginBottom="8dp"
+    app:layout_constraintBottom_toBottomOf="parent"
+    app:layout_constraintEnd_toEndOf="parent"
+    app:layout_constraintStart_toStartOf="parent"
+    app:layout_constraintTop_toTopOf="parent">
+
+    <android.support.design.widget.TextInputEditText
+        android:id="@+id/TextInputEditText"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:hint="Input Name" />
+</android.support.design.widget.TextInputLayout>
+
+<Button
+    android:id="@+id/buttonInputName"
+    android:layout_width="0dp"
+    android:layout_height="52dp"
+    android:layout_marginTop="48dp"
+    android:text="Войти"
+    app:layout_constraintEnd_toEndOf="@+id/textInputLayout"
+    app:layout_constraintHorizontal_bias="0.0"
+    app:layout_constraintStart_toStartOf="@+id/textInputLayout"
+    app:layout_constraintTop_toBottomOf="@+id/textInputLayout" />
+
+<Button
+    style="?attr/borderlessButtonStyle"
+    android:id="@+id/buttonClose"
+    android:layout_width="0dp"
+    android:layout_height="wrap_content"
+    android:layout_marginTop="56dp"
+    android:text="Закрыть"
+    app:layout_constraintEnd_toEndOf="@+id/buttonInputName"
+    app:layout_constraintStart_toStartOf="@+id/buttonInputName"
+    app:layout_constraintTop_toBottomOf="@+id/buttonInputName" />
+
+</android.support.constraint.ConstraintLayout>
