+++ app/src/main/res/values/strings.xml
@@ -8,6 +8,4 @@
     <string name="action_settings">Settings</string>
     <string name="scheme2">Scheme2</string>
     <string name="scheme1">Scheme1</string>
-
-    <string name="about_title">О пользователе</string>
 </resources>
